# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RentSaathi - Indian Landlord Rent Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/{propertyId}/units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List units in a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Units list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Add a unit to a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get a single unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Add a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Unit:
      type: object
      required:
        - id
        - propertyId
        - unitNumber
        - type
        - rentAmount
        - status
        - createdAt
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        unitNumber:
          type: string
        floor:
          type: ["string", "null"]
        type:
          type: string
          enum: [room, shop, flat, pg, studio, office]
        rentAmount:
          type: number
        depositAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [occupied, vacant, maintenance]
        amenities:
          type: ["string", "null"]
        tenantId:
          type: ["integer", "null"]
        tenantName:
          type: ["string", "null"]
        rentDueDay:
          type: ["integer", "null"]
        lastPaymentDate:
          type: ["string", "null"]
        paymentStatus:
          type: ["string", "null"]
          enum: [paid, due, overdue, partial, null]
        createdAt:
          type: string
    UnitInput:
      type: object
      required:
        - unitNumber
        - type
        - rentAmount
      properties:
        unitNumber:
          type: string
        floor:
          type: string
        type:
          type: string
          enum: [room, shop, flat, pg, studio, office]
        rentAmount:
          type: number
        depositAmount:
          type: number
        amenities:
          type: string
        rentDueDay:
          type: integer
    UnitUpdate:
      type: object
      properties:
        unitNumber:
          type: string
        floor:
          type: string
        type:
          type: string
          enum: [room, shop, flat, pg, studio, office]
        rentAmount:
          type: number
        depositAmount:
          type: number
        status:
          type: string
          enum: [occupied, vacant, maintenance]
        amenities:
          type: string
        tenantId:
          type: ["integer", "null"]
        rentDueDay:
          type: integer

    # ── Tenant schemas ────────────────────────────────────────────────────────
...
    Vendor:
      type: object
      required:
        - id
        - name
        - phone
        - category
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        category:
          type: string
          enum: [plumber, electrician, carpenter, painter, pest_control, cleaner, other]
        rating:
          type: ["number", "null"]
        rateCard:
          type: ["string", "null"]
        available:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    VendorInput:
      type: object
      required:
        - name
        - phone
        - category
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        category:
          type: string
          enum: [plumber, electrician, carpenter, painter, pest_control, cleaner, other]
        rating:
          type: number
        rateCard:
          type: string
        notes:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        category:
          type: string
          enum: [plumber, electrician, carpenter, painter, pest_control, cleaner, other]
        rating:
          type: number
        rateCard:
          type: string
        available:
          type: boolean
        notes:
          type: string