# lib/api-zod/src/generated/types/unitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentSaathi - Indian Landlord Rent Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UnitInputType } from './unitInputType';

export interface UnitInput {
  unitNumber: string;
  floor?: string;
  type: UnitInputType;
  rentAmount: number;
  depositAmount?: number;
  amenities?: string;
  rentDueDay?: number;
}